refactor(audit): Replace ChangeAuditStampsHelper with ChangeAuditStampsMixin for streamlined audit stamp management in Chart and Dashboard classes
