+++ metadata-ingestion/src/datahub/sdk/_shared.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import warnings
+from abc import ABC, abstractmethod
 from datetime import datetime
 from typing import (
     TYPE_CHECKING,
@@ -111,109 +112,110 @@ def parse_time_stamp(ts: Optional[models.TimeStampClass]) -> Optional[datetime]:
     return parse_ts_millis(ts.time)
 
 
-class ChangeAuditStampsHelper:
-    """Helper class for managing audit stamps on entities."""
+class ChangeAuditStampsMixin(ABC):
+    """Mixin class for managing audit stamps on entities."""
 
-    @staticmethod
-    def get_last_modified(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[datetime]:
+    __slots__ = ()
+
+    @abstractmethod
+    def _get_audit_stamps(self) -> models.ChangeAuditStampsClass:
+        """Get the audit stamps from the entity properties."""
+        pass
+
+    @abstractmethod
+    def _set_audit_stamps(self, audit_stamps: models.ChangeAuditStampsClass) -> None:
+        """Set the audit stamps on the entity properties."""
+        pass
+
+    @property
+    def last_modified(self) -> Optional[datetime]:
         """Get the last modification timestamp from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.lastModified.time == 0:
             return None
         return datetime.fromtimestamp(
             audit_stamps.lastModified.time / 1000
         )  # supports only seconds precision
 
-    @staticmethod
-    def set_last_modified(
-        audit_stamps: models.ChangeAuditStampsClass, last_modified: datetime
-    ) -> None:
+    def set_last_modified(self, last_modified: datetime) -> None:
         """Set the last modification timestamp in audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         audit_stamps.lastModified.time = make_ts_millis(last_modified)
+        self._set_audit_stamps(audit_stamps)
 
-    @staticmethod
-    def get_last_modified_by(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[str]:
+    @property
+    def last_modified_by(self) -> Optional[str]:
         """Get the last modification actor from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.lastModified.actor == builder.UNKNOWN_USER:
             return None
         return audit_stamps.lastModified.actor
 
-    @staticmethod
-    def set_last_modified_by(
-        audit_stamps: models.ChangeAuditStampsClass, last_modified_by: ActorUrnOrStr
-    ) -> None:
+    def set_last_modified_by(self, last_modified_by: ActorUrnOrStr) -> None:
         """Set the last modification actor in audit stamps."""
         if isinstance(last_modified_by, str):
             last_modified_by = make_user_urn(last_modified_by)
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         audit_stamps.lastModified.actor = str(last_modified_by)
+        self._set_audit_stamps(audit_stamps)
 
-    @staticmethod
-    def get_created_at(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[datetime]:
+    @property
+    def created_at(self) -> Optional[datetime]:
         """Get the creation timestamp from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.created.time == 0:
             return None
         return datetime.fromtimestamp(
             audit_stamps.created.time / 1000
         )  # supports only seconds precision
 
-    @staticmethod
-    def set_created_at(
-        audit_stamps: models.ChangeAuditStampsClass, created_at: datetime
-    ) -> None:
+    def set_created_at(self, created_at: datetime) -> None:
         """Set the creation timestamp in audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         audit_stamps.created.time = make_ts_millis(created_at)
+        self._set_audit_stamps(audit_stamps)
 
-    @staticmethod
-    def get_created_by(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[ActorUrnOrStr]:
+    @property
+    def created_by(self) -> Optional[ActorUrnOrStr]:
         """Get the creation actor from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.created.actor == builder.UNKNOWN_USER:
             return None
         return audit_stamps.created.actor
 
-    @staticmethod
-    def set_created_by(
-        audit_stamps: models.ChangeAuditStampsClass, created_by: ActorUrnOrStr
-    ) -> None:
+    def set_created_by(self, created_by: ActorUrnOrStr) -> None:
         """Set the creation actor in audit stamps."""
         if isinstance(created_by, str):
             created_by = make_user_urn(created_by)
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         audit_stamps.created.actor = str(created_by)
+        self._set_audit_stamps(audit_stamps)
 
-    @staticmethod
-    def get_deleted_on(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[datetime]:
+    @property
+    def deleted_on(self) -> Optional[datetime]:
         """Get the deletion timestamp from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.deleted is None or audit_stamps.deleted.time == 0:
             return None
         return datetime.fromtimestamp(
             audit_stamps.deleted.time / 1000
         )  # supports only seconds precision
 
-    @staticmethod
-    def set_deleted_on(
-        audit_stamps: models.ChangeAuditStampsClass, deleted_on: datetime
-    ) -> None:
+    def set_deleted_on(self, deleted_on: datetime) -> None:
         """Set the deletion timestamp in audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         # Default constructor sets deleted to None
         if audit_stamps.deleted is None:
             audit_stamps.deleted = models.AuditStampClass(
                 time=0, actor=builder.UNKNOWN_USER
             )
         audit_stamps.deleted.time = make_ts_millis(deleted_on)
+        self._set_audit_stamps(audit_stamps)
 
-    @staticmethod
-    def get_deleted_by(
-        audit_stamps: models.ChangeAuditStampsClass,
-    ) -> Optional[ActorUrnOrStr]:
+    @property
+    def deleted_by(self) -> Optional[ActorUrnOrStr]:
         """Get the deletion actor from audit stamps."""
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if (
             audit_stamps.deleted is None
             or audit_stamps.deleted.actor == builder.UNKNOWN_USER
@@ -221,18 +223,17 @@ class ChangeAuditStampsHelper:
             return None
         return audit_stamps.deleted.actor
 
-    @staticmethod
-    def set_deleted_by(
-        audit_stamps: models.ChangeAuditStampsClass, deleted_by: ActorUrnOrStr
-    ) -> None:
+    def set_deleted_by(self, deleted_by: ActorUrnOrStr) -> None:
         """Set the deletion actor in audit stamps."""
         if isinstance(deleted_by, str):
             deleted_by = make_user_urn(deleted_by)
+        audit_stamps: models.ChangeAuditStampsClass = self._get_audit_stamps()
         if audit_stamps.deleted is None:
             audit_stamps.deleted = models.AuditStampClass(
                 time=0, actor=builder.UNKNOWN_USER
             )
         audit_stamps.deleted.actor = str(deleted_by)
+        self._set_audit_stamps(audit_stamps)
 
 
 class HasPlatformInstance(Entity):
